shared/requirements.txt
uvicorn
sqlalchemy
openai
pydantic
pydantic[email]
fastapi
asyncpg
python-jose
passlib
python-dotenv
argon2_cffi
kafka-python
aiokafka
orjson
